Move ConfigProvider to separate file to fix HMR
